7.2.3 bump mapping with height map and normal map in tangent/world space.

+++ Assets/03Materials/BlinnPhong.shader
@@ -26,8 +26,8 @@
 
 			struct v2f {
 				float4 pos : SV_POSITION;
-				fixed3 worldNormal : TEXCOORD0;
-				fixed3 worldPos : TEXCOORD1;
+				float3 worldNormal : TEXCOORD0;
+				float3 worldPos : TEXCOORD1;
 			};
 
 			v2f vert(a2v v) {
